+++ src/App.js
@@ -7,7 +7,7 @@ import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 import {
   AddStudent,
   // SingleCampus,
-  DisplayCampusesContainer,
+  AllCampusesContainer,
   EditStudentContainer,
   EditCampusContainer,
   AllStudentsContainer,
@@ -22,7 +22,7 @@ class App extends Component {
         <div>(Navbar goes here)</div>
         {/* It doesn't seem to recognize <navbar> */}
         <Switch>
-          <Route exact path="/campuses" component={DisplayCampusesContainer} />
+          <Route exact path="/campuses" component={AllCampusesContainer} />
           {/* <Route exact path="/student/:id" component={SingleCampus} /> */}
           {/* <Route exact path="/editcampus/:id" component={EditCampus} /> */}
           <Route
@@ -33,6 +33,7 @@ class App extends Component {
           <Route exact path="/addstudent" component={AddStudent} />
 
           {/* URLs for forms editing an individual entry will also need an id */}
+          {/* ALSO feel free to change the path */}
           <Route exact path="/editcampus" component={EditCampusContainer} />
           <Route exact path="/allstudents" component={AllStudentsContainer} />
           <Route exact path="/addcampus" component={AddCampusContainer} />
